Move & Correct translation scripts
Moved, renamed & edited the translation scripts
